on_the_porch/rag stuff/requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
openai
python-dotenv

